[pkg] Move create*Session functions into new aws_session.go
